--- src/pages/tickoPages/BasicAccountPage.jsx
@@ -1,11 +1,0 @@
-import Header from "../../components/tickoPages/common/header/Header";
-import AccountBody from "../../components/tickoPages/basicAccountPage/AccountBody";
-
-export default function BasicAccountPage() {
-  return (
-    <>
-      <Header link="/ticko"></Header>
-      <AccountBody></AccountBody>
-    </>
-  );
-}
