--- src/pages/tickoPages/BasicChartPage.jsx
@@ -1,11 +1,0 @@
-import Header from "../../components/tickoPages/common/header/Header";
-import ChartBody from "../../components/tickoPages/basicChartPage/ChartBody";
-
-export default function BasicChartPage() {
-  return (
-    <>
-      <Header link="/ticko"></Header>
-      <ChartBody></ChartBody>
-    </>
-  );
-}
