--- src/pages/tickoPages/BasicWordPage.jsx
@@ -1,11 +1,0 @@
-import Header from "../../components/tickoPages/common/header/Header";
-import WordBody from "../../components/tickoPages/basicWordPage/WordBody";
-
-export default function BasicWordPage() {
-  return (
-    <>
-      <Header link="/ticko"></Header>
-      <WordBody></WordBody>
-    </>
-  );
-}
