+++ src/pages/tickoPages/accountMenuPages/AccountMenuPage2.jsx
@@ -1,7 +1,7 @@
 import Header from "../../../components/tickoPages/common/header/Header";
 import AccountBody2 from "../../../components/tickoPages/basicAccountPage/menu/AccountBody2";
 
-export default function AccountPage2() {
+export default function AccountMenuPage2() {
   return (
     <>
       <Header link="/ticko/account"></Header>
